Fix issue in `test_rmsnorm.py` that fails on Keras 3.6.0

+++ tests/layers/normalizations/test_rmsnorm.py
@@ -1,14 +1,14 @@
 import numpy as np
+import pytest
 from einops import asnumpy, rearrange
 from keras import layers, models, ops
-import pytest
 
 from keras_mml.layers import RMSNorm
 
 
 # Calls
 def test_no_learnable_weights_1():
-    x = np.array([1, 2, 3])
+    x = np.array([1, 2, 3], dtype="float")
     y = RMSNorm(has_learnable_weights=False)(x)
     y_pred = asnumpy(y)
     y_true = np.array([0.46291004, 0.92582010, 1.38873015])
@@ -16,7 +16,7 @@ def test_no_learnable_weights_1():
 
 
 def test_no_learnable_weights_2():
-    x = np.array([4, -5, 6, -7])
+    x = np.array([4, -5, 6, -7], dtype="float")
     y = RMSNorm(has_learnable_weights=False)(x)
     y_pred = asnumpy(y)
     y_true = [0.71269665, -0.89087081, 1.06904497, -1.24721913]
@@ -24,7 +24,7 @@ def test_no_learnable_weights_2():
 
 
 def test_with_learnable_weights_1d():
-    x = np.array([1, 2, 3])
+    x = np.array([1, 2, 3], dtype="float")
     y = RMSNorm(use_bias=False)(x)
     assert ops.shape(y) == ops.shape(x)
 
@@ -33,7 +33,7 @@ def test_with_learnable_weights_1d():
 
 
 def test_with_learnable_weights_2d():
-    x = np.array([[1, 2, 3], [4, 5, 6]])
+    x = np.array([[1, 2, 3], [4, 5, 6]], dtype="float")
     y = RMSNorm(use_bias=False)(x)
     assert ops.shape(y) == ops.shape(x)
 
@@ -42,7 +42,9 @@ def test_with_learnable_weights_2d():
 
 
 def test_with_learnable_weights_3d():
-    x = np.array([[[1, 2, 3], [4, 5, 6]], [[1, 2, 3], [4, 5, 6]], [[1, 2, 3], [4, 5, 6]], [[1, 2, 3], [4, 5, 6]]])
+    x = np.array(
+        [[[1, 2, 3], [4, 5, 6]], [[1, 2, 3], [4, 5, 6]], [[1, 2, 3], [4, 5, 6]], [[1, 2, 3], [4, 5, 6]]], dtype="float"
+    )
     y = RMSNorm(use_bias=False)(x)
     assert ops.shape(y) == ops.shape(x)
 
@@ -58,7 +60,7 @@ def test_scale_access():
 # Training
 def test_training():
     # Dataset is just a sequence of known numbers
-    x = np.array([1, 2, 3, 4, 5])
+    x = np.array([1, 2, 3, 4, 5], dtype="float")
     x = rearrange(x, "(b f) -> b f", f=1)
     y = np.copy(x)
